Styled Header
Added styling for the header to all pages. All pages still validate. Added some extra functionality also e.g. JavaScript function for the date.

Some things to fix in future - a minimum size needs to be added for the logo as if you resize too small the date/buttons float on top of the logo. Also the in-page links from the nav no longer work correctly as the "start" of the content is hidden behind the fixed header.