Homepage validation errors fixed
A large no of validation errors in homepage resolved - main categories included multiple uses of IDs, multiple classes being applied to an element, giving image size in the HTML as % etc. Resolved for now, but this page will need a lot more work to strip out the inline CSS everywhere.